Added a DAG.java class
I decided to add a new class to implement the LCA for a DAG class because my original LCA.java class (master branch) used an implementation that if I were to modify it it to allow for a directed acyclic graph, I would be pretty much replacing the entire thing as I didn't use Edges/Vertices etc. originally.